trivia-database-generation/requirements.txt
mysql-connector-python==9.2.0
python-dotenv==1.0.1
schema==0.7.7
